[FEAT] plotting both, biased and unbiased range error histograms
HINT: the unbiased range error histgramm is motivated by the fact that
range errors are always positive, thus the histogram is in case of an
unbiased Gaussian error not centered around zero. A solution is to mirror
the error around 0 in order to fit a distribution. In case of an biased
error, this would inflate the actual uncertainty. Thus, both biased and
unbiased histograms are now supported.

Signed-off-by: rojung <roland.jung@aau.at>